+++ libMx/i_micro488ex.h
@@ -8,7 +8,7 @@
  *
  *------------------------------------------------------------------------
  *
- * Copyright 2004 Illinois Institute of Technology
+ * Copyright 2004, 2006 Illinois Institute of Technology
  *
  * See the file "LICENSE" for information on usage and redistribution
  * of this file, and for a DISCLAIMER OF ALL WARRANTIES.
@@ -43,31 +43,31 @@ MX_API mx_status_type mxi_micro488ex_print_interface_structure(
 					FILE *file, MX_RECORD *record );
 MX_API mx_status_type mxi_micro488ex_open( MX_RECORD *record );
 
-MX_API mx_status_type mxi_micro488ex_open_device(MX_GPIB *gpib, int address);
-MX_API mx_status_type mxi_micro488ex_close_device(MX_GPIB *gpib, int address);
+MX_API mx_status_type mxi_micro488ex_open_device(MX_GPIB *gpib, long address);
+MX_API mx_status_type mxi_micro488ex_close_device(MX_GPIB *gpib, long address);
 MX_API mx_status_type mxi_micro488ex_read( MX_GPIB *gpib,
-						int address,
+						long address,
 						char *buffer,
 						size_t max_bytes_to_read,
 						size_t *bytes_read,
-						int flags);
+						unsigned long flags);
 MX_API mx_status_type mxi_micro488ex_write( MX_GPIB *gpib,
-						int address,
+						long address,
 						char *buffer,
 						size_t bytes_to_write,
 						size_t *bytes_written,
-						int flags);
+						unsigned long flags);
 MX_API mx_status_type mxi_micro488ex_interface_clear(MX_GPIB *gpib);
 MX_API mx_status_type mxi_micro488ex_device_clear(MX_GPIB *gpib);
 MX_API mx_status_type mxi_micro488ex_selective_device_clear(MX_GPIB *gpib,
-						int address);
+						long address);
 MX_API mx_status_type mxi_micro488ex_local_lockout(MX_GPIB *gpib);
-MX_API mx_status_type mxi_micro488ex_remote_enable(MX_GPIB *gpib, int address);
-MX_API mx_status_type mxi_micro488ex_go_to_local(MX_GPIB *gpib, int address);
-MX_API mx_status_type mxi_micro488ex_trigger(MX_GPIB *gpib, int address);
+MX_API mx_status_type mxi_micro488ex_remote_enable(MX_GPIB *gpib, long address);
+MX_API mx_status_type mxi_micro488ex_go_to_local(MX_GPIB *gpib, long address);
+MX_API mx_status_type mxi_micro488ex_trigger(MX_GPIB *gpib, long address);
 MX_API mx_status_type mxi_micro488ex_wait_for_service_request( MX_GPIB *gpib,
 						double timeout );
-MX_API mx_status_type mxi_micro488ex_serial_poll( MX_GPIB *gpib, int address,
+MX_API mx_status_type mxi_micro488ex_serial_poll( MX_GPIB *gpib, long address,
 					unsigned char *serial_poll_byte );
 MX_API mx_status_type mxi_micro488ex_serial_poll_disable(MX_GPIB *gpib);
